Earlier in the song I used the term "galvanistic," and galvanism is the concept, uh, the obsolete scientific theory that there is a kind of electricity flowing through our bloodstreams, and that was our life force. I used the term because I came across it in, uh, Mary Shelley's, uh, "Frankenstein", and that book is sort of an exploration of the theme of creating a character, of making up a person. So I used the term "galvanistic" to allude to that book as a sort of a symbol of how I, like, created you as a character. I'm pretending that I know a lot more about you than I actually do, and also to refer to the fact that I've fall—fallen in love with the characters you've created in, uh, your body of work  This is the part of the song where I start to regret writing it
Earlier in the song I used the term "galvanistic," and galvanism is the concept, uh, the obsolete scientific theory that there is a kind of electricity flowing through our bloodstreams, and that was our life force. I used the term because I came across it in, uh, Mary Shelley's, uh, "Frankenstein", and that book is sort of an exploration of the theme of creating a character, of making up a person. So I used the term "galvanistic" to allude to that book as a sort of a symbol of how I, like, created you as a character. I'm pretending that I know a lot more about you than I actually do, and also to refer to the fact that I've fall—fallen in love with the characters you've created in, uh, your body of work

This is the part of the song where I start to regret writing it